PSR-4 fix and changed to Composer
