+++ composer.json
@@ -1,0 +1,12 @@
+{
+  "name": "gladosdan/bee",
+  "description": "Bee is a torrent tracker announce generator created for use by torrent tracker administrators to debug their tracker.",
+  "minimum-stability": "stable",
+  "type": "library",
+  "license": "MIT",
+  "autoload": {
+    "psr-4": {
+      "GLaDOSDan\\Bee\\": "src/"
+    }
+  }
+}
